# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem HRIS (Human Resource Information System)",
        "title": "HRIS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/absensi/clock-in": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Karyawan melakukan clock in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Clock In",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/absensi/clock-out": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Karyawan melakukan clock out",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Clock Out",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/absensi/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil riwayat absensi user yang sedang login",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Get absensi saya",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bulan (1-12)",
                        "name": "bulan",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tahun (contoh: 2026)",
                        "name": "tahun",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/absensi/today": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil semua absensi hari ini (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Get absensi hari ini",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/cuti": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil semua pengajuan cuti (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuti"
                ],
                "summary": "Get semua cuti",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter status (pending/approved/rejected)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Karyawan mengajukan permohonan cuti",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuti"
                ],
                "summary": "Ajukan cuti",
                "parameters": [
                    {
                        "description": "Cuti Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CutiRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cuti/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil riwayat cuti user yang sedang login",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuti"
                ],
                "summary": "Get cuti saya",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cuti/{id}/approve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "HR menyetujui atau menolak pengajuan cuti",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuti"
                ],
                "summary": "Approve/Reject cuti",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cuti ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Approve Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApproveCutiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.ApproveCutiRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "catatan_hr": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.CutiRequest": {
            "type": "object",
            "required": [
                "alasan",
                "tanggal_akhir",
                "tanggal_mulai",
                "tipe_cuti"
            ],
            "properties": {
                "alasan": {
                    "type": "string"
                },
                "tanggal_akhir": {
                    "type": "string"
                },
                "tanggal_mulai": {
                    "type": "string"
                },
                "tipe_cuti": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}